game_master_agent/pyproject.toml
[project]
name = "game-master-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.6.8",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
]
